Move VeteranInformation to shared
