+++ src/applications/appeals/996/pages/veteranInformation.js
@@ -1,4 +1,4 @@
-import VeteranInformation from '../components/VeteranInformation';
+import VeteranInformation from '../../shared/components/VeteranInformation';
 
 const veteranInformation = {
   uiSchema: {
